Move ticking sound to SpeakStream
