+++ src/main.rs
@@ -746,9 +746,6 @@ fn run_get_content_wait_on_file(file_path: &Path) -> Result<String, String> {
 static FAILED_TEMP_FILE: LazyLock<NamedTempFile> =
     LazyLock::new(|| create_temp_file_from_bytes(include_bytes!("../assets/failed.mp3"), ".mp3"));
 
-static TICK_TEMP_FILE: LazyLock<NamedTempFile> =
-    LazyLock::new(|| create_temp_file_from_bytes(include_bytes!("../assets/tick.mp3"), ".mp3"));
-
 /// A global, lazily-initialized closure for sending paths into a channel.
 static PLAY_AUDIO: LazyLock<Box<dyn Fn(&Path) + Send + Sync>> = LazyLock::new(|| {
     // Create a channel for path buffers.
@@ -788,46 +785,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
         Some(voice) => voice.into(),
         None => Voice::Echo,
     };
-    let (speak_stream, speak_state_rx) = ss::SpeakStream::new(ai_voice, opt.speech_speed);
+    let (speak_stream, speak_state_rx) = ss::SpeakStream::new(ai_voice, opt.speech_speed, opt.tick);
     let speak_stream_mutex = Arc::new(Mutex::new(speak_stream));
-
-    if opt.tick {
-        let state_rx = speak_state_rx.clone();
-        thread::spawn(move || {
-            let tick_sink = DefaultDeviceSink::new();
-            let tick_path = TICK_TEMP_FILE.path().to_path_buf();
-            let mut should_tick = false;
-            let mut has_spoken = false;
-            loop {
-                match state_rx.recv_timeout(Duration::from_millis(100)) {
-                    Ok(ss::SpeakState::Converting) => {
-                        if !has_spoken {
-                            should_tick = true;
-                        }
-                    }
-                    Ok(ss::SpeakState::Playing) => {
-                        has_spoken = true;
-                        should_tick = false;
-                        tick_sink.stop();
-                    }
-                    Ok(ss::SpeakState::Idle) => {
-                        has_spoken = false;
-                        should_tick = false;
-                        tick_sink.stop();
-                    }
-                    Err(flume::RecvTimeoutError::Disconnected) => break,
-                    Err(flume::RecvTimeoutError::Timeout) => {}
-                }
-
-                if should_tick && tick_sink.empty() {
-                    if let Ok(file) = std::fs::File::open(&tick_path) {
-                        tick_sink.stop();
-                        tick_sink.append(rodio::Decoder::new(BufReader::new(file)).unwrap());
-                    }
-                }
-            }
-        });
-    }
 
     match opt.subcommands {
         Some(subcommand) => {
